# Faster time encoding for Handle Scribe