--- a/katip/test/Main.hs
+++ b/katip/test/Main.hs
@@ -4,6 +4,7 @@ module Main (main) where
 import           Test.Tasty
 -------------------------------------------------------------------------------
 import qualified Katip.Tests
+import qualified Katip.Tests.Format.Time
 import qualified Katip.Tests.Scribes.Handle
 -------------------------------------------------------------------------------
 
@@ -15,5 +16,6 @@ testSuite :: TestTree
 testSuite = testGroup "katip"
   [
     Katip.Tests.tests
+  , Katip.Tests.Format.Time.tests
   , Katip.Tests.Scribes.Handle.tests
   ]