This implements time encoding functions using low-level Data.Test interface.

Encoding functions do not allocate anything except the result, and work really fast.

The Handle Scribe benchmark improved almost 4x times. From 24us to 6.1 us.

This should probably be used in JSON encoding, at least in the toJSON part.

The toEncoding part should use efficient implementation in Data.Aeson to avoid allocating Text or ByteString outside of Builder.

Fixes #3

Benchmark results before

```
benchmarking Katip.Scribes.Handle/Bytestring Builder
time                 23.85 μs   (23.41 μs .. 24.35 μs)
                     0.996 R²   (0.993 R² .. 0.999 R²)
mean                 24.11 μs   (23.68 μs .. 24.82 μs)
std dev              1.978 μs   (1.220 μs .. 2.994 μs)
variance introduced by outliers: 79% (severely inflated)
```

After

```
benchmarking Katip.Scribes.Handle/Bytestring Builder
time                 6.106 μs   (6.025 μs .. 6.223 μs)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 6.212 μs   (6.099 μs .. 6.524 μs)
std dev              569.2 ns   (258.0 ns .. 1.164 μs)
variance introduced by outliers: 85% (severely inflated)
```

We should add JSON benchmark and check if this may improve it also.