@MichaelXavier 2016-10-26
@ozataman Any thoughts on this? The benchmark results look pretty nice. This change would not break any compatibility as its an internal change. Currently I am not aware of any plans to support custom time formats in the Handle scribe (of course it could be accomplished with a custom scribe), and ISO8601 seems like the best target format. I am a bit surprised there isn't a package out there that provides this formatting functionality. Maybe there is but it targets a different time library.