@ozataman 2016-10-26
@sopvop Thank you for this work! I'm certainly not opposed to performance improvements, especially when they don't change anything at the interface level for the user. 

Lacking the time to go into it myself right now, here are the few things I'd want to ask to make sure:

- Currently, the output time format wouldn't change at all, right? I.e. no visible changes can be discerned by the user after this update?

- Any partial applications or anything like that that could raise unexpected exceptions due to this change?

- I've noticed there's already an associated test suite, which is great. Since this change relies on custom-made rendering functions, I guess it may be a good idea to take a quick look and see if we've covered all obvious cases. Common pitfalls in date related rendering typically are things like 0-padding output dates, etc. 

In general, our biggest fear is to be a source of exceptions or missed logs in production for the users of our library. Thanks again for the work @sopvop!