@sopvop 2016-10-26
Btw, bikeshedding is very much welcome. If you don't like function or module names or anything.

I've added the ISO8601 variant so it can be used in json encoding part. Also both of these functions may be useful for implementing other scribes for katip.

--------

Not directly related to this PR:

Another possible improvement to Handle Scribe is encoding to utf-8 lazy bytestring before outpting it to handle.

Text.hPutBuf goes through GHC locale handling and iconv, even though most of computers use utf-8 based locales, at least on unixes.

Encoding to utf8 and using `D.B.Lazy.hPut` bench improves even more, shaving off another second.

```
benchmarking Katip.Scribes.Handle/Bytestring Builder
time                 4.840 μs   (4.775 μs .. 4.937 μs)
                     0.997 R²   (0.994 R² .. 0.999 R²)
mean                 4.891 μs   (4.823 μs .. 5.148 μs)
std dev              384.2 ns   (197.7 ns .. 757.9 ns)
variance introduced by outliers: 81% (severely inflated)
```
Changing existing handler like that is breaking BC, so we may instead introduce mkHandleScribeUtf8.

And in that case changing from Text.Builder to ByteString.Builder may improve things even more...