@sopvop 2016-10-26
It should output exactly same format as before. I've added tests for that case.
The only difference should be years before 1000, those will be zero padded. And negative years will be absed.

There are no partial functions, the only thin place is allocating Array of proper length. And I've double-checked that. (you may want to tripple-check that, just to be sure)

Also this patch does not change the elasticsearch scribe.