@sopvop 2016-10-31
>  Currently, the output time format wouldn't change at all, right? I.e. no visible changes can be discerned by the user after this update?

In last commit I've added test against "golden" output by Handle scribe generated by original code.

>  Any partial applications or anything like that that could raise unexpected exceptions due to this change?

All functions are total, unlike formatter from aeson, which is partial with negative years.

>   I've noticed there's already an associated test suite, which is great. Since this change relies on custom-made rendering functions, I guess it may be a good idea to take a quick look and see if we've covered all obvious cases. Common pitfalls in date related rendering typically are things like 0-padding output dates, etc.

I've added several tests to check that output matches, if you think I've missed some corner cases, I'll be happy to add tests for them.

The only possible formatting change is zero padded years, and I don't think this is relevant for a logger.