@MichaelXavier 2016-10-31
@sopvop Thanks for your hard work on this! Everything looks great. I'll probably be cutting a new release today. I'll add your name to the contributors list.