@MichaelXavier 2016-10-31
Released in 0.3.1.0.